# application/src/main/java/com/jobshunter/dto/JobOrderRequest.java
package com.jobshunter.dto;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

@JsonInclude(JsonInclude.Include.NON_NULL)
@JsonIgnoreProperties(ignoreUnknown = true)
public record JobOrderRequest(
    @NotNull @Positive Long engineId,
    boolean searchCompanies,
    boolean searchWithUserPrompts
) {
}