# application/src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: JobsHunter API
  version: "1.0.0"
  description: |
    HTTP API for user authentication, job search, profile management, CV upload, and admin email tools.
servers:
  - url: https://ronin-archesporial-lonnie.ngrok-free.dev
    description: Prod
  - url: http://localhost:8081
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]
    RegistrationResponse:
      type: object
      properties:
        message:
          type: string
        verificationToken:
          type: string
      required: [message, verificationToken]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 255
        phoneNumber:
          type: string
          maxLength: 30
      required: [username, email, password, phoneNumber]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
...
    EngineType:
      type: string
      enum: [GPT, GEMINI, SERP]
...
    UploadCvResponse:
      type: object
      properties:
        message:
          type: string
        fileId:
          type: string
      required: [message, fileId]
...
    ModelResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Engine configuration ID
        model:
          type: string
          description: Model name
      required: [id, model]
    JobOrderRequest:
      type: object
      properties:
        engineId:
          type: integer
          format: int64
          description: Engine configuration ID
        searchCompanies:
          type: boolean
          description: Whether to search for companies
        searchWithUserPrompts:
          type: boolean
          description: Whether to use user prompts for the search
      required: [engineId, searchCompanies, searchWithUserPrompts]
    JobOrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Job order ID
        engineConfigurationId:
          type: integer
          format: int64
          description: Engine configuration ID
        engine:
          type: string
          description: Engine type (GPT, GEMINI, SERP)
        model:
          type: string
          description: Model name
        searchCompanies:
          type: boolean
          description: Whether to search for companies
        status:
          $ref: '#/components/schemas/OrderStatus'
        notified:
          type: boolean
          description: Whether the user has been notified
        timestamp:
          type: string
          format: date-time
          description: Order creation timestamp
        errorMessage:
          type: string
          nullable: true
          description: Error message if the order failed
      required: [id, engineConfigurationId, engine, model, searchCompanies, status, notified, timestamp]
    OrderStatus:
      type: string
      enum: [NEW, PROCESSING, COMPLETED, FAILED, CANCELED]
      description: Status of the job order
...
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered and verification token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      summary: Obtain JWT token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/verify:
    patch:
      summary: Verify user email by token
      tags: [Auth]
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
            maxLength: 128
      responses:
        '200':
          description: Email verified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Token not found or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cv/upload:
    post:
      summary: Upload user CV to ChatGPT
      tags: [CV]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '200':
          description: CV uploaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadCvResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cv:
    delete:
      summary: Delete user CV
      tags: [CV]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CV deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/models:
    get:
      summary: Get available models for an engine
      tags: [Models]
      description: Returns list of all available models for the specified engine type. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: engine
          required: true
          schema:
            $ref: '#/components/schemas/EngineType'
          description: Engine type to filter models by
      responses:
        '200':
          description: List of models for the specified engine
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelResponse'
        '400':
          description: Invalid engine parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/engine/models:
    get:
      summary: Get all engine models grouped by engine
      tags: [Engine]
      description: Returns a map where keys are engine types and values are lists of available models for each engine. Requires authentication.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Map of engines to their available models
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
                example:
                  GPT:
                    - gpt-4o-mini
                    - gpt-5.2
                  GEMINI:
                    - gemini-2.5-flash
                    - gemini-2.5-pro
                  SERP:
                    - google_jobs
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/engine/order:
    post:
      summary: Create a new job order
      tags: [Engine]
      description: Creates a new job order for the authenticated user with the specified engine configuration. Requires authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobOrderRequest'
      responses:
        '201':
          description: Job order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    description: Job order ID
                  message:
                    type: string
                    description: Success message
                required: [id, message]
        '400':
          description: Invalid request (e.g., engine configuration not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/engine/orders:
    get:
      summary: Get job orders for authenticated user
      tags: [Engine]
      description: Returns all job orders for the authenticated user, sorted by timestamp (descending) and then by status. Requires authentication.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of job orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobOrderResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/engine:
    get:
      summary: Get models for a specific engine
      tags: [Engine]
      description: Returns a list of model names available for the specified engine type. Requires authentication.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: engine
          required: true
          schema:
            $ref: '#/components/schemas/EngineType'
          description: Engine type to get models for
      responses:
        '200':
          description: List of model names for the specified engine
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - gpt-4o-mini
                  - gpt-5.2
        '400':
          description: Invalid engine parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...